Unify windows snapshotters & add support to store scratch layers in different directory (#27)
* Refactor windows snapshotters

Both LCOW & WCOW snapshotters have many functions that are identical and some other
functions have common code. This change refactors those snapshotters to remove duplicate
code.
This refactor will also help when adding other features like storing scratch snapshots in a
different location.

* Support to override scratch snapshot location

Some containers do heavy IO to the scratch, in case of LCOW/WCOW the scratch itself is a
VHD that is stored on the same disk as that of other snapshots (including the image
layers). If containerd is pulling multiple images and one or more containers are doing
heavy IO to the scratch at the same time, that causes disk contention and related
issues. This change allows the user to set a config (either in containerd.toml or in the
container configuration) that specifies another directory inside which the scratch of all
containers (or that specific container) should be stored.

(cherry picked from commit 333a72bfcd4831549d074d824164d45ef9b4d0db)
